agent/requirements.txt
# Python 3.10+ dependencies
# Core ML/AI
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0

# LLM/VLM Models
qwen-vl-utils>=0.0.1
dashscope>=1.14.0
openai>=1.3.0
anthropic>=0.8.0

# Vector Database
chromadb>=0.4.18
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Agent Framework
langchain>=0.1.0
langgraph>=0.0.20

# Audio Processing
openai-whisper>=20231117
librosa>=0.10.1
soundfile>=0.12.1

# Image/Video Processing
Pillow>=10.1.0
opencv-python>=4.8.1
numpy>=1.24.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.25.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
redis>=5.0.1

# Frontend Build
npm>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.2
tqdm>=4.66.1
PyJWT>=2.8.0

# API Clients
requests>=2.31.0
tencentcloud-sdk>=3.0.1009

# Document Processing
python-docx>=1.1.0
reportlab>=4.0.7
